[kong]: Refactor and merge small crates into kong
