--- krates/kdata/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "kdata"
-version.workspace = true
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-krypto = { path = "../krypto/"}
-serde.workspace = true
-serde_json.workspace = true
-chrono.workspace = true
