--- krates/kdata/README.md
@@ -1,5 +1,0 @@
----
-title: kdata
-subtitle: Common data structures used in kong
-...
----
