--- krates/kdata/src/property.rs
@@ -1,1 +1,0 @@
-
