--- krates/kerror/README.md
@@ -1,5 +1,0 @@
----
-title: kerror
-subtitle: Common error types used in kong
-...
----
