--- krates/konfig/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "konfig"
-version.workspace = true
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-kerror = { path = "../kerror/"}
-serde.workspace = true
-toml.workspace = true
