--- krates/konfig/README.md
@@ -1,5 +1,0 @@
----
-title: konfig
-subtitle: kong server configuration
-...
----
