--- krates/konfig/src/defaults.rs
@@ -1,8 +1,0 @@
-//! `kong` default values
-
-/// Kong server working directory
-pub const WORKING_DIRECTORY: &str = "~/kong/";
-/// Kong server databases working directory
-pub const DBS_DIRECTORY: &str = "databases/";
-/// Accounts directory inside the working directory
-pub const ACCONTS_DB: &str = "databases/ACCOUNTS.sqlite";
