+++ krates/kong/Cargo.toml
@@ -9,11 +9,10 @@ edition = "2021"
 database = []
 
 [dependencies]
-kdata = { path = "../kdata/"}
-konfig = { path = "../konfig/"}
 krypto = { path = "../krypto/"}
-kerror = { path = "../kerror"}
 rouille.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 route-recognizer.workspace = true
+chrono.workspace = true
+toml.workspace = true
