+++ krates/kong/src/error.rs
@@ -1,0 +1,24 @@
+use std::fmt;
+
+#[derive(Debug)]
+/// Kong errors
+pub enum KError {
+    /// Configuration error
+    Config,
+    /// URL parsing error
+    UrlParsing,
+    /// Invalid HTTP Method
+    InvalidHttpMethod,
+}
+
+impl std::error::Error for KError {}
+
+impl fmt::Display for KError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Config => write!(f, "Could not read config file"),
+            Self::UrlParsing => write!(f, "Could not parse error"),
+            Self::InvalidHttpMethod => write!(f, "Invalid HTTP method"),
+        }
+    }
+}
