--- krates/kong/src/prelude.rs
@@ -1,16 +1,0 @@
-//! Re-exports of upstream crates
-
-pub use crate::kontrol::{
-    error_response::ErrorResponse, issue_kpassport::is_admin, Kontrol, Method,
-};
-pub use crate::kroute::kroute;
-pub use crate::Kong;
-pub use kdata;
-pub use kerror::KError;
-pub use konfig::Konfig;
-pub use krypto;
-pub use rouille as server;
-pub use serde_json;
-pub use serde_json::error::Error as JsonError;
-pub use serde_json::json;
-pub use serde_json::Value as JsonValue;
