+++ krates/kong/src/validate.rs
@@ -5,6 +5,7 @@
 //! ## References
 //! - <https://beesbuzz.biz/code/439-Falsehoods-programmers-believe-about-email>
 
+/// User input validator
 pub struct Validate;
 
 impl Validate {
@@ -74,17 +75,22 @@ impl Validate {
 use std::fmt;
 
 #[derive(Debug)]
+/// Validation error
 pub enum ValidationError {
+    /// Generic validation error
     Generic,
+    /// Username validation error
     Username,
+    /// Email validation error
     Email,
+    /// Password validation error
     Password,
 }
 
 impl std::error::Error for ValidationError {}
 
 impl fmt::Display for ValidationError {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
             ValidationError::Generic => write!(f, "General input error"),
             ValidationError::Username => write!(f, "Invalid username"),
