+++ krates/krypto/Cargo.toml
@@ -11,4 +11,3 @@ blake3.workspace = true
 chrono.workspace = true
 base64.workspace = true
 hex.workspace = true
-kerror = { path = "../kerror/"}
