+++ krates/krypto/src/defaults.rs
@@ -1,3 +1,5 @@
+//! Default values
+
 /// The length of a username
 /// A username cannot be longer than 15 characters.
 /// A username can only contain alphanumeric characters (letters A-Z,
